Resolve whether parameter types are actually type variables.
Parameters can either be a declaration of a generic type, or a
statement about the concrete type the parameter is taking for a
particular instance. These are difficult to determine at initial
parse, so resolve whether the class exists once parsing is finished
and assume that the parameter is a type variable if it does not.

Bug: 31858628
Test: run update-api, android.signature.cts.SignatureTest#testSignature
Change-Id: I88d316cece8640fd5d356482ebbe0426d918d8f4