Offer to emit exact APIs instead of superset.
When using the "showAnnotation" option, we generate a superset of
public API plus any annotated methods.  This is useful for normal
stubs JAR generation, but it makes it hard to compare only the diffs
of a specific annotation type.

To remedy this, offer new "-exactApi" output file which will emit
exactly the APIs annotated with the "-showAnnotation" value.  This
was done by refactoring the existing "-removedApi" logic to support
any sort of filtering we want to perform.

Test: make update-api && make -j32 online-system-api-sdk-docs
Bug: 38146029
Change-Id: Ice51814db8daa0214ab91b5fa088eed0dd69410a