Add -stubsourceonly option
Add option to cause Doclava to only generate stubs for classes whose
source is provided. This makes it possible to create stubs for a subset
of all the classes in a package.

This is needed for generating stubs for the JUnit and related classes
being removed from the Android API which includes
com.android.internal.util.Predicate.

Bug: 30188076
Test: make update-legacy-test-api
Change-Id: Icc7355cb562072f5604b6c25c4d620002924317c