Add -stubimportpackages
The stub creation checks to ensure that the generated stubs are self
consistent and does not reference classes that are not themselves in the
stubs. This change modifies that to allow certain packages to be treated
as if they are imported into the stubs, i.e. as part of a runtime
dependency. The new -stubimportpackages option takes a : separated list
of packages that should be treated as imports. Specifying a package in
that has two effects:
* It hides all classes in that package so that they are not added to the
  stubs.
* It does not report an error if stub classes reference classes in that
  package.

Bug: 30188076
Test: make checkbuild
Change-Id: I945fc8a199d6bebe25de17c79c2bf65d8e0e6b8d