Add -convert2xmlnostrip option
The apicheck -convert2xml option will not generate <implements> elements
for interfaces that are not included in the api description being
converted. That works fine when the input is a complete api description
such as frameworks/base/api/current.txt but causes problems when the
input describes an optional module, such as
frameworks/base/legacy-test/api/legacy-test-current.txt.

This adds an extra -convert2xmlnostrip option that will not do any
stripping.

Tested by using it to convert both files mentioned above. Verified that
it had no impact on the api generated for current.txt and that it added
the missing <implements> elements.

Bug: 35192974
Test: see above
Change-Id: I3d20c22ab40e7524f8c5050ce4b0e9610768dc7c