Correctly stub classes with hidden constructors
If a class had no publicly-visible constructors, and the
super constructor had a throws clause, no call to a super
constructor would be generated. If there was no default
super constructor, this would cause a build failure.

Fixed by just allowing the super call even though the throws
are not exposed.

Bug: 67777702
Test: Stubs that failed to build now build.
Change-Id: I7d1126daefee097091ec7c96d9b85694db7f2f8f