Emit all APIs using Predicates.
Awhile back we wrote a mechanism to emit API based on a customizable
Predicate<MemberInfo> which would "filter" exactly what should be
included or excluded.

This change extends this mechanism to be used for all API emission,
specifically "current.txt" files.  To accomplish this, we now have
two predicates: one for deciding if a member should be emitted
directly into the API being generated, and another to decide if a
member can be referenced indirectly as part of emitting another API.

For example, generating removed.txt, we only want to emit members
tagged with "@removed", but of course it's okay for those members to
reference a non-removed class like String, etc.

Rewrite ApiPredicate to be general enough to support various matching
requirements.  Also search package hierarchy when deciding if a
member matches or not, since "@hide" is often defined at the package
level.

Only emit superclasses and interfaces that match our predicate.  If
a superclass/interface doesn't match the predicate, walk further
up the family tree to look for any and all superclasses/interfaces
that do actually match our predicate.  Consider this example:

    public class StringBuilder extends AbstractStringBuilder {}
    class AbstractStringBuilder implements Appendable {}

In this case, we want to say "StringBuilder implements Appendable",
because AbstractStringBuilder won't be emitted by our predicate.

For fields inherited from ancestors, if a field would have matched had
it been defined directly on the class being inspected, then clone that
field to make it match.  For methods inherited from ancestors, don't
allow the class being inspected to pretend that their override is
hidden.

Split out eliding logic into a separate predicate that is applied
as one final step.  Fix bug where we weren't emitting generics.

Test: manual inspection of API text files
Bug: 69505783
Change-Id: Ic7f349c6cb80853967308d2f871a0b029ead0096
Merged-In: Ic7f349c6cb80853967308d2f871a0b029ead0096