Add -showUnannotated
Previously, the -showAnnotation option generated an API definition file
that included all the unannotated APIs as well as any API annotated with
the specified annotation. This was changed to make it only show the APIs
that were explicitly annotated with the specified annotation and to
exclude the unannotated APIs.

See change Ic7f349c6cb80853967308d2f871a0b029ead0096 for more details.

Unfortunately, that broke some uses that were made of the API definition
file. So, this adds a new option -showUnannotated that when specified
will cause the generated API definition files to include all the
unannotated APIs even when -showAnnotation is also specified.

Follow on change will use this and test the new behavior.

Bug: 71420101
Bug: 30188076
Test: make update-api to ensure it does not change any existing APIs
Change-Id: I9b803b33fd22599b910db7aabd8c41471a76d2a7