Refactor code to make predicates negatable
Previous refactoring of the stubs generator moved most of the API
membership logic into Predicate classes. Bits of the method filtering
logic, however, remained hard-coded in the API printer. This prevents
usage of more complicated predicates, namely predicates filtering
members that are *not* present in the public API. This patch refactors
the code further to enable these use cases.

Test: dbrazdil@ checked that generated files have not changed
Change-Id: I8f329de5154d2e16614361990f7615cb5f80ea9e