Add option to print the complement of public API
Enforcing access control for non-SDK APIs requires a list of classes,
methods and fields which are not present in the SDK. This patch adds
the option to generate an ApiFile which contains the complement set
of the public SDK stubs.

This particular file will be passed to the Signature CTS test to
check that private APIs are not discoverable.

Test: manual inspection of generated files
Bug: 64382372
Change-Id: I995b87f6f71ba75ae77ac27f2a8e01b9e8f41e07