Add option to generate DEX signatures of private API
Enforcing access control no APIs outside the public SDK requires marking
such classes, methods and fields in the DEX files of the boot class
path. This patch adds a new set of writer methods which emit a list of
DEX-level signatures of members which match a predicate, and uses it
to emit a list of DEX signatures for private APIs.

Test: manual inspecition of generated files
Bug: 64382372
Change-Id: Ia018a5e1bd428d98e63177e905ed113298d3792f