Pick private APIs from all classes
Printing private APIs would use the wrong set of packages which do
not include @hide classes or classes implementing an interface but
never referenced directly. These classes should be included in the
private API lists and this patch will pass the complete class list
to the writers.

Note that the public API filter is now applied on non-strippable
classes, while its negation (the private API filter) is applied on
all classes. This is fine as we do not want to consider the
strippable classes hidden.

Test: manual
Bug: 64382372
Change-Id: Iacdb6bc776a837528a8a18588e16e50c7cae3184