Escape space chars in anchor tags, remove legacy <a> tags for anchors, and cleanup legacy DAC markup.
The non-escaped spaces have been a long-standing issue that created
incompatible behaviors on the new devsite design. Escaping these is the
right thing to do, but yes, it will break some existing links that use
the old URLs with actual spaces. We'll need to resolve those over time.

Using the <a> tags resulted in bad vertical positioning, so this now puts the
anchor name in the header tag with the id attribute.

Also lots of removal of legacy "apilevel" class names, and addition
of the data attribute where it was still missing.

bug: 79995820 (escape spaces)
bug: 80256515 (remove "apilevel" classes)

test: make ds-docs
sample of results at cl/197951140

Change-Id: I038cd4abfdd498f82740aeb149cbaabd252d3ff1
Merged-In: I038cd4abfdd498f82740aeb149cbaabd252d3ff1