Fix bad class literals in generated annotations
When generating annotated stubs code for members with public
annotations, if there is a class literal in the annotation the
generated code will not compile.

For example:

@Foo(myClass=a.b.c.D.class)
void myMethod() {}

...becomes...

@Foo(myClass=a.b.c.D)
void myMethod() {}

... which is not valid Java.

There are no current examples of this in Android but it was noticed
while adding a new hidden annotation that didn't yet have @hide.

Test: Add a public annotation with a class literal, build, fails without
      this change.

Change-Id: Ib3eea66cab6d94948237a8c458b092b17ed993e6