Make broken @see tags errors in doclava outside frameworks.
Prior to this change, doclava reported errors for @link tags that
cross-referenced to something that didn't exist or wasn't visible, but
had an exemption for @see tags.

Recent changes[1] have removed all such broken @see tags outside of
frameworks. This change removes the exemption, and so detects
regressions. (Broken @see tags in frameworks are fixed downstream, and
there is a change downstream to remove the exemption globally. The
latter change is used as the merged-in for this change. That will mean
we catch regressions downstream, but that's less convenient than
catching them directly in AOSP for libcore, external/icu, etc.)

This is the second attempt at this change. The first attempt[2] caused
issues downstream, with violations that don't show up in AOSP. This
attempt has been tested in an appropriate downstream branch.

[1] See:
https://android-review.googlesource.com/c/platform/libcore/+/704719
https://android-review.googlesource.com/c/platform/external/icu/+/704721
https://android-review.googlesource.com/c/platform/libcore/+/705144

[2] See:
https://android-review.googlesource.com/c/platform/external/doclava/+/705225

Bug: 80570421
Test: make docs
Test: make docs, cherry-picked into pi-dev
Change-Id: I92852fab0a8d22b115fa89ce4302abe1ca082bd6
Merged-In: I5ad574025d5e2709a853a82d2d52124526520c69