Support for dex signature -> source position map.
For every class processed by doclava, output the dex signature and
source position of the corresponding API definition. This can then be
read by other tools to perform automated changes to the source file.

A new parameter -apiMapping is added to control this functionality. The
mapping file is in the format of:
  dex-signature
  source/file.java:123

with two lines for every dex signature.

Test: m
Bug: 110868826

Change-Id: I12e01419e10efb7f38706f1a24d8bdf44f5bf55e