Automatic documentation for "@Column" annotation.
We have various "contract" classes across the OS which describe the
columns that are available for insert(), update(), and query() when
working with ContentProviders.

To help ensure that the underlying providers fully support the API
contracts being made, this change defines a new "@Column" annotation
which describes the underlying data type.

These annotations can then be used by the ContentProvider
implementations internally to ensure that they're fully meeting the
API contracts being made.

Bug: 120429729
Test: manual
Change-Id: Iffeb5fce4a1186e0b4aa3424ecb69a34f6607c21