Update doclava to use @since tags from comments.
Metalava is now generating the API stubs and handles the calculation
of when APIs were introduced and deprecated. It now directly adds this
info to each APIs comment using the tags @apiSince and
@deprecatedSince.

This cl enables the doclava API reference generator to extract these
tags and output them to the docs. Requires passing -metalavaApiSince
into the doclava build.

bug: 125939989
Test: make ds-docs

Change-Id: I1a6c04a3e05dd1dc74ccdaf81ec45b52a40f55fc