Add workaround for androidx prebuilt references
androidx generates some of their docs against prebuilts, which
can't be updated in sync with the doclava changes to inner class
constructor references. This change adds a flag that is off by default
that will enable us to conditionally use the old workaround for
referring to inner class contructor references.

This was an attempt to scope the workaround to the smallest possible
applications to the workaround.

Once we have updated prebuilts, this workaround can be removed.

Bug: 6963924
Test: ./gradlew publicDocsTask and ./gradlew disttipOfTreeDocs
Change-Id: I87098a7497e6de51b93baae97a8ca8fdefc6129d
Merged-In: I87098a7497e6de51b93baae97a8ca8fdefc6129d